create shared folder for reusable actions, actionCreatores and constants
