--- src/client/actions.ts
@@ -1,10 +1,0 @@
-import { Action } from 'redux';
-
-export interface DefaultAction extends Action {
-  type: ''
-}
-
-export interface PayloadAction<P> extends Action {
-  type: string;
-  payload?: P;
-}
