+++ src/client/containers/App/reducer.ts
@@ -1,7 +1,7 @@
 import { Reducer } from 'redux';
 import * as types from './constants';
 import { FeatureAction } from './actions';
-import { defaultAction } from 'actionCreators';
+import { defaultAction } from 'shared/actionCreators';
 import { State, initialState } from './state';
 
 const reducer: Reducer<State> = (state: State = initialState, action: FeatureAction = defaultAction()) => {
