+++ src/client/shared/actionCreators.ts
@@ -1,0 +1,6 @@
+import {DefaultAction} from './actions';
+import * as types from './constants';
+
+export const defaultAction = (): DefaultAction => ({
+  type: types.DEFAULT_ACTION_TYPE,
+});
