+++ src/client/shared/constants.ts
@@ -1,0 +1,2 @@
+export type DEFAULT_ACTION_TYPE = '';
+export const DEFAULT_ACTION_TYPE: DEFAULT_ACTION_TYPE = '';
