+++ tsconfig.json
@@ -13,8 +13,7 @@
     "noUnusedLocals": true,
     "paths": {
       "components/*": ["./components/*"],
-      "containers/*": ["./containers/*"],
-      "actionCreators": ["./actionCreators"]      
+      "containers/*": ["./containers/*"]
     }
   },
   "include": [
